requirements.txt
pytest==8.3.5
playwright==1.50.0
python-dotenv==1.0.1